back-end/requirements.txt
# requirements.txt

# Core dependencies
fastapi>=0.100.0,<0.101.0
uvicorn>=0.22.0,<0.23.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.27.0,<0.28.0
psycopg2-binary>=2.9.5,<3.0.0
alembic>=1.11.0,<2.0.0

# Authentication
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
bcrypt>=4.0.1,<5.0.0
python-multipart>=0.0.5,<0.1.0

# Blockchain
web3>=6.0.0,<7.0.0
substrate-interface>=1.2.0,<2.0.0
eth-account>=0.8.0,<0.9.0

# Utils
python-dotenv>=1.0.0,<2.0.0
requests>=2.28.0,<3.0.0
httpx>=0.24.0,<0.25.0
tenacity>=8.2.0,<9.0.0
loguru>=0.7.0,<0.8.0
aiofiles>=23.1.0,<24.0.0
pytz>=2023.3

# Testing
pytest>=7.3.1,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0